requirements.txt
Django==4.2.11
djangorestframework==3.15.0
model_bakery==1.17.0
tqdm==4.66.2
drf_yasg==1.21.7
